backend/requirements.txt
# BRANE Backend Dependencies
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication
authlib==1.3.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.12
httpx==0.27.2

# LLM Integration
litellm==1.48.7
sentence-transformers==3.1.1
openai==1.51.2
anthropic==0.39.0

# Vector Store (FAISS)
faiss-cpu==1.8.0.post1
numpy==1.26.4

# Security & Privacy
cryptography==43.0.1
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355

# Utilities
pyyaml==6.0.2
python-dotenv==1.0.1
redis==5.1.1
aiofiles==24.1.0
PyPDF2==3.0.1

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
httpx==0.27.2
